chore: Move files around to flatten hierarchy
