+++ packages/core/src/renderer/components/dock/edit-resource/edit-resource-model/call-for-resource.global-override-for-injectable.ts
